Add files via upload
Reaper lacks the option to set default velocity on midi notes. This is a workaround, that lets you control "preset" velocity from a jsfx slider, that can be linked to a control surface.
It also has a primitive note preview system.  The script manages the loading of the jsfx files.